+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - **net/http/pprof:** `RegisterRoutes` to allow registering routes in other routers than the `http.DefaultServeMux`
 - **x/net/trace:** `RegisterRoutes` to allow registering routes in other routers than the `http.DefaultServeMux`
 - **x/net/trace:** `NoAuth` to allow remote tracing
+- **expvar:** `RegisterRoutes` to allow registering routes in other routers than the `http.DefaultServeMux`
 
 
 ## 0.7.0 - 2017-07-27